feat(std/io): add node runtime and buffer type for handle and param in open and exists
