+++ packages/std/src/io/runtime/bun/exists.ts
@@ -1,24 +1,13 @@
-import { fileURLToPath } from 'node:url'
-import { exists as existsDefinition, FSError, handle as handleDefinition, type Impl, IOErrors, Runtime } from 'std:io'
-import { guard, isFailure } from 'std:result'
-import { isString } from 'std:shared'
-import { stats as nodeStatsDefinition } from '../node/stats'
+import { exists as existsDefinition, type FSError, type Impl, IOErrors, Runtime } from 'std:io'
+import { guard } from 'std:result'
+import { exists as nodeExistsDefinition } from '../node/exists'
 
 export const exists = existsDefinition.extend(({ use }): Impl.Exists<FSError | IOErrors.exists> => {
-  const statsApi = use(nodeStatsDefinition)
-  const handleApi = use(handleDefinition)
+  const nodeExistsApi = use(nodeExistsDefinition)
 
   return guard(
-    async path => {
-      const resolvedHandle = path instanceof URL ? fileURLToPath(path) : path
-      const handle = isString(resolvedHandle) ? handleApi(resolvedHandle) : resolvedHandle
-      const result = await statsApi.stats(handle)
-      if (isFailure(result)) {
-        if (result.error instanceof FSError && result.error.code === 'ENOENT') {
-          return false
-        }
-      }
-      return handle.assembled
+    async function* (path) {
+      return yield* await nodeExistsApi(path)
     },
     IOErrors.exists,
     Runtime.bun,
