+++ packages/std/src/io/runtime/node/index.ts
@@ -1,12 +1,12 @@
 import { extendableIO } from 'std:io'
 import { createPlugin } from 'std:plugin'
-
+import { exists } from './exists'
 import { open } from './open'
 import { read } from './read'
 import { stats } from './stats'
 import { write } from './write'
 
-export const createNodeIO = createPlugin(extendableIO.clone().define(stats, open, read, write), {
+export const createNodeIO = createPlugin(extendableIO.clone().define(stats, open, read, write, exists), {
   name: 'node',
 })
